Move grid logic into separate components
