+++ src/components/GridPlot/GridAxis.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export interface Props {
+  children: React.ReactChild | React.ReactChild[];
+}
+
+export const GridXAxis: React.FC<Props> = () => {
+  return null;
+};
+
+export const GridYAxis: React.FC<Props> = () => {
+  return null;
+};
